update the path variable

+++ src/main/java/org/springframework/samples/petclinic/developer/DeveloperUiController.java
@@ -70,7 +70,7 @@ public class DeveloperUiController {
 	}
 
 	@GetMapping("/developers/delete/{id}")
-	public String deleteDeveloper(@PathVariable int id) {
+	public String deleteDeveloper(@PathVariable Long id) {
 
 		this.developerService.deleteById(id);
